pyproject.toml
[project]
name = "spacetraders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
